simplify main.ts
