+++ release/app/package.json
@@ -1,12 +1,12 @@
 {
-  "name": "electron-react-boilerplate",
-  "version": "4.6.0",
-  "description": "A foundation for scalable desktop apps",
+  "name": "noteapp",
+  "version": "0.1.0",
+  "description": "note-taking app",
   "license": "MIT",
   "author": {
-    "name": "Electron React Boilerplate Maintainers",
-    "email": "electronreactboilerplate@gmail.com",
-    "url": "https://github.com/electron-react-boilerplate"
+    "name": "Sophie Bernadin-Mercier",
+    "email": "sophiebmercier@hotmail.com",
+    "url": "https://github.com/Shuhala"
   },
   "main": "./dist/main/main.js",
   "scripts": {
