+++ src/main/application/index.ts
@@ -1,0 +1,1 @@
+export { NoteApplication } from './noteApplication';
