+++ src/main/application/noteApplication.ts
@@ -1,18 +1,29 @@
-import { IpcMainConnection, IpcMainConnectionEvent } from './ipcMainConnection';
-import { IpcChannel, IpcChannelMessage } from '../common/ipc';
-import { ConfigurationService } from './configuration/configurationService';
+import {
+  IpcMainConnection,
+  IpcMainConnectionEvent,
+} from '../ipcMainConnection';
+import { IpcChannel, IpcChannelMessage } from '../../common/ipc';
+import { ConfigurationService } from '../configuration/configurationService';
+import { initApp } from './startup/startup';
 
 /*
  * NoteApplication contains all of the application's state and logic.
  */
 export class NoteApplication {
+  electronApp: Electron.App;
+
   configurationService: ConfigurationService;
 
   constructor(electronApp: Electron.App, ipcMainConnection: IpcMainConnection) {
+    this.electronApp = electronApp;
     this.configurationService = new ConfigurationService(
       electronApp.getPath('userData')
     );
     ipcMainConnection.onPing((e, m) => this.onPing(e, m));
+  }
+
+  start(): void {
+    initApp(this.electronApp);
   }
 
   private async onPing(
