+++ src/main/application/startup/appUpdater.ts
@@ -1,0 +1,10 @@
+import { autoUpdater } from 'electron-updater';
+import log from 'electron-log';
+
+export class AppUpdater {
+  constructor() {
+    log.transports.file.level = 'info';
+    autoUpdater.logger = log;
+    autoUpdater.checkForUpdatesAndNotify();
+  }
+}
