+++ src/main/application/startup/menu.ts
