+++ src/utils.ts
@@ -1,0 +1,1 @@
+export const srcPath = __dirname;
